+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - refactor(core): use python-fake for faking
 - refactor(tests): wait for wifi status icons in `test_wireless`
 - refactor(core): move `ubo_app.utils.loop` to `ubo_app.service` and add service properties like `service_id`, `name`, `label`, etc as module level variables to it
+- refactor(core): make `UboModuleLoader` to keep a weakref of the module in its cache instead of an actual reference
 
 ## Version 0.15.4